Reorganize project's directory structure
